Time: 62 ms (38.17%) | Memory: 55.8 MB (67.47%) - LeetSync

+++ 2665-minimum-time-to-repair-cars/minimum-time-to-repair-cars.java
@@ -4,7 +4,7 @@ class Solution {
         long l = 0, r = cars * mn * cars ;
         while(l <= r){
             long md = l + (r - l)/2 ;
-            long cnt = 0 ;
+            int cnt = 0 ;
             for(int i : ranks){
                 cnt += Math.sqrt(md/i);
             }
